add math and rename ds2 nnet
